molly-2/molly-scraper/requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
extruct==0.18.0
flexcache==0.3
flexparser==0.4
grpcio==1.75.1
grpcio-tools==1.75.1
html-text==0.7.1
html5lib==1.1
idna==3.11
ingredient_parser_nlp==2.3.0
isodate==0.7.2
joblib==1.5.2
jstyleson==0.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
mf2py==2.0.1
nltk==3.9.2
numpy==2.2.6
Pint==0.24.4
platformdirs==4.5.0
protobuf==6.32.1
pyparsing==3.2.5
pyRdfa3==3.6.4
python-crfsuite==0.9.11
rdflib==7.2.1
recipe_scrapers==15.9.0
regex==2025.9.18
requests==2.32.5
six==1.17.0
soupsieve==2.8
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
w3lib==2.3.1
webencodings==0.5.1
